sys/shell: replaced old shell with new
