+++ sys/shell/pkg.yml
@@ -19,15 +19,13 @@
 
 pkg.name: sys/shell 
 pkg.description: Command processor for console-based applications.
-pkg.author: "Apache Mynewt <dev@mynewt.incubator.apache.org>"
-pkg.homepage: "http://mynewt.apache.org/"
+pkg.author: "Michal Narajowski <michal.narajowski@codecoup.pl>"
+pkg.homepage: "http://www.codecoup.pl/"
 pkg.keywords:
 
 pkg.deps:
     - kernel/os
-    - encoding/base64 
     - time/datetime
-    - util/crc
 pkg.req_apis:
     - console
