+++ sys/shell/syscfg.yml
@@ -22,9 +22,21 @@ syscfg.defs:
     SHELL_TASK:
         description: 'Controls whether shell is enabled or not.'
         value: 0
-    SHELL_MAX_INPUT_LEN:
-        description: 'Maximum input line length'
-        value: 256
-    SHELL_DEBUG:
-        description: Enables additional error checking in the shell package.
-        value: 0
+
+    SHELL_TASK_PRIO:
+        description: 'Shell task priority'
+        value: 3
+    SHELL_STACK_SIZE:
+        description: 'Shell task stack size'
+        value: 512
+
+    SHELL_CMD_HELP:
+        description: 'Include help information for shell commands'
+        value: 1
+
+    SHELL_OS_MODULE:
+        description: 'Include shell os module'
+        value: 1
+    SHELL_PROMPT_MODULE:
+        description: 'Include shell prompt module'
+        value: 1
